requirements.txt
numpy
pillow
pandas
matplotlib
colorama
geopandas
shapely
pyproj
fiona
pytest
geodatasets
